requirements.txt
flask
qrcode
pillow
fpdf2
pyopenssl
gunicorn
